requirements.txt
Flask
Flask-Session
mysql-connector-python
Flask-Mail
bcrypt
Flask-Bcrypt
python-dotenv
